requirements.txt
# =============================
# FastAPI and Server
# =============================
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# =============================
# LangGraph and LLM Ecosystem
# =============================
langgraph==1.0.2
langchain==1.0.4
langchain-core==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
litellm==1.79.1
openai==2.7.1
anthropic==0.72.0

# =============================
# MCP (Model Context Protocol)
# =============================
mcp==1.21.1

# =============================
# Database
# =============================
sqlalchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.1
psycopg2-binary==2.9.11
greenlet==3.2.4

# =============================
# Redis
# =============================
redis==7.0.1
hiredis==3.3.0

# =============================
# Authentication
# =============================
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1

# =============================
# HTTP Clients
# =============================
httpx==0.28.1
aiohttp==3.13.2

# =============================
# Observability
# =============================
langfuse==3.9.1
# OpenTelemetry core packages (required by Langfuse, but we don't use OTLP export)
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp==1.38.0

# =============================
# Data Validation
# =============================
pydantic==2.12.4
pydantic-settings==2.11.0
email-validator==2.2.0
pyyaml==6.0.3

# =============================
# Testing
# =============================
pytest==8.4.2
pytest-asyncio==1.2.0
httpx-sse==0.4.3

# =============================
# Code Quality
# =============================
ruff==0.14.4
black==25.9.0

# =============================
# Logging
# =============================
colorama==0.4.6

# =============================
# Background Jobs
# =============================
apscheduler==3.10.4

# =============================
# Travel APIs
# =============================
amadeus==12.0.0

# =============================
# TOON Format (Token-efficient LLM data format)
# =============================
toon-python==0.1.2
